I wrote a "Guess the flag!" app as part of the HackingWithSwift: 100 days of SwiftUI course, made by Paul Hudson.
The app consists of a "start menu" screen, from where a user can start a game. And the main screen in which a user can play the game.

The game consists of 8 rounds, where a user can choose the right flag from 3 different flags. The display of which is randomized.

If the user guessed right, one point will be added to the score count. If the user guessed wrong, an alert will provide the user which countries' flag they selected, and the score count will be reset to 0.

After 8 rounds, an alert will pop up, providing the user with the final score count. The alert will give the option to end the game. This will result in the user going back to the main menu.